% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ScriptP04_MakeLogregTable_v08.5_AllSetsUcorr_20141207h18m20_41k_kloci.R
\name{cnvGSAgsTables}
\alias{cnvGSAgsTables}
\title{Creating the gene-set tables for each gene-set}
\usage{
cnvGSAgsTables(cnvGSA.in, cnvGSA.out)
}
\arguments{
\item{cnvGSA.in}{A CnvGSAInput S4 object.}

\item{cnvGSA.out}{A CnvGSAOutput S4 object.}
}
\value{
A list where each object is a table corresponding to on gene-set.
}
\description{
Creates the gene-set tables for each gene-set
}

