% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ScriptP04_MakeLogregTable_v08.5_AllSetsUcorr_20141207h18m20_41k_kloci.R
\name{cnvGSAIn}
\alias{cnvGSAIn}
\title{Creating the input S4 object needed to run the script}
\usage{
cnvGSAIn(configFile)
}
\arguments{
\item{configFile}{The file name for the config file including the full path.}
}
\value{
A cnvGSAInput S4 object
}
\description{
Creating the input S4 object needed to run the script
}

