% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cnvGSA.R
\name{cnvGSAlogRegTest}
\alias{cnvGSAlogRegTest}
\title{Performing the logistic regression tests on the CNV data}
\usage{
cnvGSAlogRegTest(cnvGSA.in, cnvGSA.out)
}
\arguments{
\item{cnvGSA.in}{A CnvGSAInput S4 object.}

\item{cnvGSA.out}{A CnvGSAOutput S4 object.}
}
\value{
A list of one or two objects depending on whether or not the user includes the known loci in the analysis. Each object in the list contains the regression results for each gene set.
}
\description{
This test uses 4 different correction models and is based on a case control study. It looks at odds ratios and calculates p-values which the researcher can analyze
}

