% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/enrFiles.R
\name{f.enrFiles}
\alias{f.enrFiles}
\title{Prepare the files for the enrichment maps.}
\usage{
f.enrFiles(cnvGSA.in, cnvGSA.out)
}
\arguments{
\item{cnvGSA.in}{A CnvGSAInput S4 object}

\item{cnvGSA.out}{A CnvGSAOutput S4 object}
}
\value{
Returns a list with the data frames of the GMT file and the generic file.
}
\description{
Prepare the files for the enrichment maps.
}
\examples{
## See vignette for full details and worked example
}

