% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cnvGSA.R
\name{f.readConfig}
\alias{f.readConfig}
\title{Reading in the config file}
\usage{
f.readConfig(configFile, cnvGSA.in)
}
\arguments{
\item{configFile}{The file name and full path for the config file}

\item{cnvGSA.in}{A CnvGSAInput S4 object.}
}
\value{
A CnvGSAInput object with the updated config.ls and params.ls objects.
}
\description{
This function is used to read in all the values from the config file and change them in the S4 objects used throughout the scripts. If you would like to reload the config values, you will want to run this function.
}
\examples{
library(cnvGSAdata)
data(cnvGSA_input_example)
## See vignette for full details and worked example
}

