% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cnvGSA.R
\name{cnvGSAgsTables}
\alias{cnvGSAgsTables}
\title{Creating the gene-set tables for each gene-set}
\usage{
cnvGSAgsTables(cnvGSA.in, cnvGSA.out)
}
\arguments{
\item{cnvGSA.in}{A CnvGSAInput S4 object.}

\item{cnvGSA.out}{A CnvGSAOutput S4 object.}
}
\value{
A list where each object is a table corresponding to on gene-set.
}
\description{
Creates the gene-set tables for each gene-set
}
\examples{
library(cnvGSAdata)
data(cnvGSA_input_example)
data(cnvGSA_output_example)
cnvGSA.out <- cnvGSAgsTables(cnvGSA.in,cnvGSA.out)
## See vignette for full details and worked example
}

