% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cnvGSA.R
\name{cnvGSAIn}
\alias{cnvGSAIn}
\title{Creating the input S4 object needed to run the script}
\usage{
cnvGSAIn(configFile, cnvGSA.in)
}
\arguments{
\item{configFile}{The file name for the config file including the full path.}

\item{cnvGSA.in}{A CnvGSAInput S4 object.}
}
\value{
A cnvGSAInput S4 object
}
\description{
Creating the input S4 object needed to run the script
}

