% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ScriptP03_Viz_20150108h09m10.R
\name{f.makeViz}
\alias{f.makeViz}
\title{Creating the plots from the CnvGSAOutput data}
\usage{
f.makeViz(cnvGSA.in, cnvGSA.out)
}
\arguments{
\item{cnvGSA.in}{A CnvGSAInput S4 object}

\item{cnvGSA.out}{A CnvGSAOutput S4 object}
}
\value{
Creates the plots to better understand the output
}
\description{
Creating the plots from the CnvGSAOutput data
}

